requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openpyxl==3.1.2
openai==1.12.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0 